refactor(core): Split workflow review request service (no-changelog)
